wip: replace catalog collection request modal through more generic implementation
