+++ apps/frontend-manage/src/components/catalog/CatalogBrowser.tsx
@@ -3,6 +3,8 @@ import { GetCatalogCollectionInfoDocument } from '@klicker-uzh/graphql/dist/ops'
 import Loader from '@klicker-uzh/shared-components/src/Loader'
 import { useRouter } from 'next/router'
 import { useState } from 'react'
+import ObjectImport from './actions/ObjectImport'
+import PendingSharingRequests from './actions/PendingSharingRequests'
 import AddObjectToCatalogButton from './administration/AddObjectToCatalogButton'
 import AddObjectToCatalogModal from './administration/AddObjectToCatalogModal'
 import ObjectAddedErrorToast from './administration/ObjectAddedErrorToast'
@@ -11,8 +13,6 @@ import CatalogCollectionCreationErrorToast from './collections/CatalogCollection
 import CatalogCollectionCreationSuccessToast from './collections/CatalogCollectionCreationSuccessToast'
 import CreateCatalogCollectionButton from './collections/CreateCatalogCollectionButton'
 import CreateCatalogCollectionModal from './collections/CreateCatalogCollectionModal'
-import ObjectImport from './import/ObjectImport'
-import PendingSharingRequests from './import/PendingSharingRequests'
 
 function CatalogBrowser() {
   const router = useRouter()
