+++ apps/frontend-manage/src/components/catalog/actions/ObjectChangeAccessModal.tsx
