+++ apps/frontend-manage/src/components/catalog/actions/ObjectRemovalModal.tsx
