+++ apps/frontend-manage/src/components/catalog/actions/ObjectSharingModal.tsx
