+++ apps/frontend-manage/src/components/catalog/actions/RequestAnswerCollectionModal.tsx
@@ -2,16 +2,19 @@ import { useMutation, useQuery } from '@apollo/client'
 import { faPaperPlane } from '@fortawesome/free-regular-svg-icons'
 import { faBan } from '@fortawesome/free-solid-svg-icons'
 import {
+  GetCatalogObjectsDocument,
   GetSingleAnswerCollectionCatalogDocument,
-  ImportAnswerCollectionDocument,
+  ObjectAccess,
+  RequestAnswerCollectionDocument,
 } from '@klicker-uzh/graphql/dist/ops'
 import { Markdown } from '@klicker-uzh/markdown'
 import Loader from '@klicker-uzh/shared-components/src/Loader'
-import { Button, Modal, Toast } from '@uzh-bf/design-system'
+import { Button, Modal, Toast, UserNotification } from '@uzh-bf/design-system'
 import { useTranslations } from 'next-intl'
 import { useState } from 'react'
 
-function ImportAnswerCollectionModal({
+// TODO: generalize this modal and only insert small part with customization depending on the object type
+function RequestAnswerCollectionModal({
   id,
   open,
   catalogCollectionId,
@@ -26,7 +29,6 @@ function ImportAnswerCollectionModal({
 }) {
   const t = useTranslations()
   const [showError, setShowError] = useState(false)
-  const [showEntries, setShowEntries] = useState(false)
 
   // fetch answer collection
   const { data, loading } = useQuery(GetSingleAnswerCollectionCatalogDocument, {
@@ -36,10 +38,45 @@ function ImportAnswerCollectionModal({
     },
   })
 
-  const [importAnswerCollection, { loading: importLoading }] = useMutation(
-    ImportAnswerCollectionDocument,
+  const [requestAnswerCollection, { loading: requestLoading }] = useMutation(
+    RequestAnswerCollectionDocument,
     {
       variables: { collectionId: id, catalogCollectionId },
+      update: (cache, { data }) => {
+        // check if request was successful
+        const requestedCollection = data?.requestAnswerCollection
+        if (!requestedCollection) return
+
+        // update lists of answer collections
+        const catalogObjects = cache.readQuery({
+          query: GetCatalogObjectsDocument,
+          variables: {
+            catalogCollectionId,
+          },
+        })
+
+        if (catalogObjects?.getCatalogObjects) {
+          const updatedObjects = catalogObjects?.getCatalogObjects.map(
+            (obj) => {
+              if (obj.id === id) {
+                return requestedCollection
+              }
+
+              return obj
+            }
+          )
+
+          cache.writeQuery({
+            query: GetCatalogObjectsDocument,
+            variables: {
+              catalogCollectionId,
+            },
+            data: {
+              getCatalogObjects: updatedObjects,
+            },
+          })
+        }
+      },
     }
   )
 
@@ -51,68 +88,40 @@ function ImportAnswerCollectionModal({
       onClose={(e) => {
         e?.stopPropagation()
         onClose()
-        setShowEntries(false)
       }}
-      title={t('manage.catalog.importPublicResource')}
+      title={t('manage.catalog.requestAccess')}
       className={{ content: 'text-sm' }}
-      dataCloseButton={{ cy: 'close-answer-collection-import-modal' }}
     >
       {loading || !collection ? (
         <Loader />
       ) : (
         <>
+          {collection.objectAccess === ObjectAccess.Public ? (
+            <UserNotification
+              type="warning"
+              message={t('manage.catalog.requestPublicResource')}
+              className={{ root: 'mb-2' }}
+            />
+          ) : null}
           <div>
-            {t.rich('manage.resources.importCollectionMessage', {
+            {t.rich('manage.resources.requestAccessMessage', {
               name: collection.name,
               owner: collection.ownerShortname,
               b: (text) => <b>{text}</b>,
             })}
           </div>
-          <div className="border-uzh-grey-100 mt-2 rounded border border-solid bg-slate-100 p-2">
+          <div className="border-uzh-grey-100 mt-2 rounded border border-solid p-2">
             <Markdown
               content={`**Description:** ${collection.description}`}
-              data={{ cy: 'import-answer-collection-description' }}
+              data={{ cy: 'request-answer-collection-description' }}
             />
-            <div>
-              {showEntries ? (
-                <div className="mt-2">
-                  <div className="font-bold">
-                    {t('manage.resources.answerOptions')}
-                  </div>
-                  <ul className="list-inside list-disc">
-                    {collection.entries?.map((entry, ix) => (
-                      <li
-                        key={entry.id}
-                        data-cy={`public-collection-answer-option-${ix}`}
-                      >
-                        {entry.value}
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-              ) : (
-                <Button
-                  basic
-                  className={{
-                    root: 'text-primary-100 px-0 py-0 hover:bg-transparent',
-                  }}
-                  onClick={() => setShowEntries(true)}
-                  data={{ cy: 'public-collection-show-answers' }}
-                >
-                  <Button.Label>
-                    {t('manage.resources.showAnswers')}
-                  </Button.Label>
-                </Button>
-              )}
-            </div>
           </div>
           <div className="mt-3 flex flex-row justify-between">
             <Button
               className={{ root: 'h-8 border-red-600 py-0' }}
-              data={{ cy: 'cancel-answer-collection-import' }}
+              data={{ cy: 'cancel-answer-collection-request' }}
               onClick={(e) => {
                 e?.stopPropagation()
-                setShowEntries(false)
                 onClose()
               }}
             >
@@ -124,28 +133,26 @@ function ImportAnswerCollectionModal({
               className={{ root: 'h-8 py-0' }}
               onClick={async (e) => {
                 e?.stopPropagation()
-                const res = await importAnswerCollection()
-                if (res.data?.importAnswerCollection) {
+                const res = await requestAnswerCollection()
+                if (res.data?.requestAnswerCollection) {
                   onSuccess()
                   onClose()
                 } else {
                   setShowError(true)
                 }
               }}
-              loading={importLoading}
-              data={{ cy: 'confirm-answer-collection-import' }}
+              loading={requestLoading}
+              data={{ cy: 'confirm-answer-collection-request' }}
             >
               <Button.Icon icon={faPaperPlane} />
-              <Button.Label>
-                {t('manage.resources.importCollection')}
-              </Button.Label>
+              <Button.Label>{t('manage.resources.requestAccess')}</Button.Label>
             </Button>
           </div>
           <Toast
             openExternal={showError}
             onCloseExternal={() => setShowError(false)}
           >
-            {t('manage.resources.importError')}
+            {t('manage.resources.requestError')}
           </Toast>
         </>
       )}
@@ -153,4 +160,4 @@ function ImportAnswerCollectionModal({
   )
 }
 
-export default ImportAnswerCollectionModal
+export default RequestAnswerCollectionModal
