+++ apps/frontend-manage/src/components/catalog/actions/SharingRequestApprovalModal.tsx
