+++ packages/graphql/src/schema/resource.ts
@@ -60,7 +60,7 @@ export const AnswerCollection = AnswerCollectionRef.implement({
   }),
 })
 
-// TODO: replace this type with something generic and generic sharing modals (should be the same for everything)
+// TODO: remove this type once the modified and generalized setup is in place
 interface ICatalogAnswerCollection extends DB.AnswerCollection {
   objectAccess: DB.ObjectAccess
   ownerShortname?: string
